+++ backend/migrations/versions/e6f1a8b25c71_change_email_inbox.py
@@ -41,7 +41,7 @@ def upgrade() -> None:
         sa.Column(
             "created_at",
             sa.DateTime(timezone=True),
-            server_default=sa.text("now()"),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
             nullable=False,
         ),
         sa.Column("workspace_id", sa.Uuid(), nullable=False),
@@ -76,7 +76,7 @@ def upgrade() -> None:
         sa.Column(
             "created_at",
             sa.DateTime(timezone=True),
-            server_default=sa.text("now()"),
+            server_default=sa.text("(CURRENT_TIMESTAMP)"),
             nullable=False,
         ),
         sa.Column("workspace_id", sa.Uuid(), nullable=False),